\name{wheat.A}
\alias{wheat.A}
\title{Pedigree info for the wheat dataset}
\description{

Is a numerator relationship matrix (599 x 599) computed
from a pedigree that traced back many generations.
This relationship matrix was derived using the
Browse application of the International Crop Information
System (ICIS), as described in  
\url{http://cropwiki.irri.org/icis/index.php/TDM_GMS_Browse} (McLaren \emph{et al.} 2005).

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
McLaren, C. G., R. Bruskiewich, A.M. Portugal, and A.B. Cosico. 2005. The International Rice Information System. 
A platform for meta-analysis of rice crop data. \emph{Plant Physiology} \bold{139}: 637-642.
}
\keyword{datasets}
